requirements/base.txt
ctx==0.1.2
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-admin-bootstrapped==2.3.5
django-ajax-selects==1.3.5
django-bootstrap-datetimepicker==1.5
django-bootstrap3-datetimepicker==2.2.3
django-braces==1.4.0
django-extensions==1.5.1
django-formset-js==0.5.0
django-jquery-js==2.1.4
django-suit==0.2.12
django-toolbelt==0.0.1
gunicorn==19.3.0
html5lib==0.999
Pillow==2.7.0
pisa==3.0.33
psycopg2==2.6.1
pydot==1.0.2
pygraphviz==1.2
pyparsing==1.5.7
reportlab==3.1.44
six==1.9.0
static3==0.6.1
Unipath==1.1
